docker_agents/a2a_agent/requirements.txt
python-dotenv
smarta2a==0.4.18
